+++ angular_router/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 2.0.0-alpha+10
+
 ### Breaking changes
 
 *   The minimum SDK version is now `sdk: ">=2.0.0-dev.46.0 <2.0.0"`.